ultima actualizacion al 05 de Octubre

+++ src/main/java/co/edu/unbosque/NamelessBack/api/ClientesAPI.java
@@ -32,7 +32,7 @@ public class ClientesAPI {
 	}
 
 	@DeleteMapping("/eliminar/{id}")
-	public void eliminar(@PathVariable("id") Integer id) {
+	public void eliminar(@PathVariable("id") Long id) {
 		clientesDAO.deleteById(id);
 	}
