+++ src/main/java/co/edu/unbosque/NamelessBack/dao/DetalleVentasDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import co.edu.unbosque.NamelessBack.model.DetalleVentas;
 
-public interface DetalleVentasDAO extends JpaRepository<DetalleVentas, Integer>{
+public interface DetalleVentasDAO extends JpaRepository<DetalleVentas, Long>{
 
 }
