+++ src/main/java/co/edu/unbosque/NamelessBack/dao/ProductosDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import co.edu.unbosque.NamelessBack.model.Productos;
 
-public interface ProductosDAO extends JpaRepository<Productos, Integer>{
+public interface ProductosDAO extends JpaRepository<Productos, Long>{
 
 }
