+++ src/main/java/co/edu/unbosque/NamelessBack/dao/ProveedoresDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import co.edu.unbosque.NamelessBack.model.Proveedores;
 
-public interface ProveedoresDAO extends JpaRepository<Proveedores, Integer>{
+public interface ProveedoresDAO extends JpaRepository<Proveedores, Long>{
 
 }
