+++ src/main/java/co/edu/unbosque/NamelessBack/dao/UsuariosDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import co.edu.unbosque.NamelessBack.model.Usuarios;
 
-public interface UsuariosDAO extends JpaRepository<Usuarios, Integer>{
+public interface UsuariosDAO extends JpaRepository<Usuarios, Long>{
 
 }
