+++ src/main/java/co/edu/unbosque/NamelessBack/dao/VentasDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import co.edu.unbosque.NamelessBack.model.Ventas;
 
-public interface VentasDAO extends JpaRepository<Ventas, Integer>{
+public interface VentasDAO extends JpaRepository<Ventas, Long>{
 
 }
